backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
celery[redis]==5.4.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
httpx==0.27.2
redis==5.1.1
neo4j==5.25.0
qdrant-client==1.12.1
pgvector==0.3.5
ollama==0.3.3
google-generativeai==0.8.3
groq==0.11.0
python-telegram-bot[webhooks]==21.6
notion-client==2.2.1
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
watchdog==5.0.3
aiofiles==24.1.0
